Introduce the main Relation interface.
As explained in Relation's documentation we introduce here an abstract
interface for Relation, over the following representations:
* in-memory relations, with the Set-based implementation simply renamed.
* lazy-evaluated relations, through relation algebra and Operator.
* on-disk relations and others, not yet implemented.

The interface provides the public API of Jalf with respect to the OO-styled
relational algebra. The functional DSL simply delegates to the OO API.

The Relation interface is not intended to be implemented directly as the
compiler chain needs access to some concrete internals. AbstractRelation
is the entry point for implementers; the compiler will cast to that abstract
class in an unchecked way.