Implement Relation#equals
As all Relation(s) denote values (even if some of them are lazy evaluated)
we want equals() to apply to those values independently of their actual
representation. This commit implements this on the basic infrastructure,
making the acceptance test pass for now.

The following logic is currently implemented:

1. AbstractRelation checks trivial cases then delegates to the abstract
   `equals(Relation)` method.
2. SetMemoryRelation implements two strategies:
   2.1 against another set-based (through underlying collection equality)
   2.2 against an arbitrary stream (through tuple inclusion).
3. Operator implements two strategies too:
   3.1 delegates to the in-memory `other` relation if it is one
   3.2 loads itself into memory to make such a delegate call.

The current strategy is not optimal at all, even if it basically works.
Possible optimizations include

1. loading the smallest relation into memory in strategy 3.2) above.
2. implement comparison of ordered streams instead of 3.2.
3. avoid the delegation complexity by extracting equals responsibility
   to another abstraction (?)